Stop saving hint number to file
When the splash window was destroyed, a random hint number was saved
to a file to be loaded the next time the splash window was displayed.
We can duplicate this functionality by just loading a random hint number
when the splash window is displayed, saving us from having to do
unnecessary disk reads and writes.