Optimize json serialization
Instead of building up the whole object tree and serializing that, just
serialize the data on the fly.

This helps reduce memory usage and speed up the serialization routines.

This new serialization code changes the output format slightly.  The
trailing .0 has been removed from floating-point values with no
fractions, and the bye member of cups is now serialized to an empty
array rather than null.