live_game: Fix bug while selecting injured player
This fixes a bug in live_game_injury_get_player() where it
would fail to chooose a player in some very rare cases.  The
function was converting the result of math_rnd() from double to
a float which caused the result to be outside the range of
values expected by the function.

The failure to choose an injured player would lead to a segfault
later on in live_game_event_injury().