Use stable pointers for Cups
Prior to this change, memory for Cups was allocated in GArray objects,
which meant that a pointer to a Cup was not stable, because the Cup
could be copied to a new location when the GArray was resized.

With this change, we allocate memory for Cups with malloc and store
Cups in GPtrArrays.