Use stable pointers for Leagues
Prior to this change, memory for Leagues was allocated in GArray objects,
which meant that a pointer to a League was not stable, because the
League could be copied to a new location when the GArray was resized.

With this change, we allocate memory for Leagues with malloc and store
Leagues in GPtrArrays.