callback_show_tables() include inactive leagues when getting next clid
There is a filter later in the function for removing inactive leagues,
so we don't need to filter them out in the
league_cup_get_{next,previous}_clid calls.

I think in the long-term it would best to remove the count_inactive
parameter from this functions, and this change makes this easier
by removing another caller that passes FALSE for count_inactive.