Fix bug while assigning game attendence
We were passing an expression with a call to rand() to the MIN macro.
Expressions passed to MIN can be evaluated twice, first for the MIN
comparision and then again for the MIN assignment.  With the rand()
call returning different values each time this meant that the value
used for the comparision could be different than the final result used
for assignment.  This could in theory lead to a game that had attendance
greater than the capacity for the teams's stadium.

In practice, though, we likely never hit this bug, because there is
another bug that causes league_cup_average_capacity() to return the
total capacity of all the stadiums in the league rather than the
average.  This means that the stadium capacity is always significantly
smaller than the value returned by league_cup_average_capacity(), so
the expression with rand() is never executed twice.