Remove clid usage from team_get_fixture()
This removes a small optimization that was skipping some
fixture lookups if the team was a 'cup team' (meaning a team
not in the current country).  I don't think this will have a
big impact on performance and will allow us eventually remove the
clid member of the Team struct.