Fix incorrect sort order in player lists
In the player lists, the number of goals show for a player is the
number of goals they have scored in whatever league/cup they will
play in next.  However, when the list is sorted, it looks at how
many goals they have scored in the league or cup with the clid
that is associated with their team.  This led to incorrect sort
order when sorting by goals, games, or shots.