Always display the team's league name in the player lists
We were displaying the cup name for teams outside of the current
country, except when the team was in the job list, then we would
display the league name.  Because of this, it seems to me like the
intention was to display the league name, it just wasn't available
until we added the league member in
be22604626e2a7322f5bb93c9016f2ab60cbf49b.