Remove av_talent arg from team_generate_players_stadium
We don't really need this any more since all teams are generated
at the start of the game now.  This should not change behavior at
all since all callers where either passing their leauge's average talent
or 0 (which would cause the function to default to using the league's
average talent), so in either case the behavior of the function was the same.