Make name assignment reproducible
This changes how names are randomly generate to ensure that the
results will be the same as long as the random seed is the same.

It also does away with shortening the name lists, which should
give the players a greater variety of names.  The shortening was
done to save memory, but I don't think the memory usage of the
lists is all that much overall.